initial implementation of multicore handler
objects are being passed by registering message handlers and sending
the objects using simple ipc methods. But the cluster implementation
uses JSON serialization in this context, so this is not suitable for
all kinds of complex objects. Furthermore plain JSON cannot deal with
class methods, hence workarounds using functions and object
parameters were necessary to get this example to run.