fix: incorrect job passed to subsequent workers
the listener to the fork event was just registered once, hence it may
have cached the first job object instead of using a new one.
Now i am not using this listener anymore, but instead pass the job
directly to the worker object which is returned after the fork() call.